requirements.txt
Django==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
docker==7.1.0
gunicorn==21.2.0
djangorestframework==3.14.0
pytest==7.4.4
pytest-django==4.7.0
pytest-mock==3.12.0
